==> nbs/pipeline/02.training.ipynb <==
import pandas as pd

df = pd.read_parquet("data/intermediate/m5_scaled.parquet")

from ts.models.nbeats import NBeatsG

horizon = 180  # <-- FORECAST HORIZON
input_size = horizon * 5


model = NBeatsG(
    input_size=input_size,
    horizon=horizon,
)

import os

import torch

from ts.preprocess.bigdata.dataloader import TSPreprocessor, UnivariateTSDataModule

batch_size = 32
step_size = 7


preprocessor = TSPreprocessor(
    df=df,
    input_size=input_size,
    horizon=horizon,
    target_col="y_scaled",
    train_split=0.7,
    val_split=0.15,
    split_type="vertical",
    step_size=step_size,
    cache_dir="cache/",
    use_cache=True,
    experiment_name="m5_experiment",
    num_workers=os.cpu_count() - 2,
    chunk_size=1000,
    adaptive_step=False,
    downsample_factor=1,
    save_by_window_count=True,
    max_windows_per_file=3000,
    batch_size=batch_size,
)

ds = UnivariateTSDataModule(
    preprocessor=preprocessor,
    batch_size=batch_size,
    num_workers=os.cpu_count(),
    pin_memory=True,
    prefetch_factor=2,
    persistent_workers=True,
    gpu_preload=False,
    experiment_name="m5_experiment",
)

import pytorch_lightning as pl
from pytorch_lightning.callbacks import EarlyStopping
from pytorch_lightning.loggers import TensorBoardLogger

checkpoint_callback = pl.callbacks.ModelCheckpoint(
    save_top_k=1,
    monitor="val_smape",
    mode="min",
    dirpath="checkpoints/",
    filename="m5_experiment",
)

trainer = pl.Trainer(
    logger=TensorBoardLogger("logs", name="m5_experiment"),
    max_epochs=200,  # Short run for testing
    accelerator="auto",
    precision="16-mixed",
    gradient_clip_val=1.0,
    callbacks=[EarlyStopping("val_smape", patience=10, verbose=False), checkpoint_callback],
    # profiler=profiler,
    accumulate_grad_batches=4,
    # log_every_n_steps=1
    # strategy="ddp_notebook"
    # strategy="ddp_notebook",
    # devices=1,
    # num_nodes=2,
)

trainer.fit(model, ds)

print(trainer.test(model, ds))
print(trainer.validate(model, ds))
