==> nbs/pipeline/04.inference.ipynb <==
from ts.models.nbeats import NBeatsG

import numpy as np
import pandas as pd
import plotly.express as px
import torch

# Device setup
device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
# Load model from checkpoint
model = NBeatsG.load_from_checkpoint("checkpoints/nbeat_m3_run1.ckpt")
model = model.to(device)
model.eval()

df = pd.read_parquet("data/intermediate/m5_inference_preprocessed.parquet")
X_scaled = df["X"].values

# Step 3: Convert to tensor
X_tensor = torch.tensor(np.stack(X_scaled), dtype=torch.float32, device=device)

# Step 4: Inference
with torch.no_grad():
    yhat_tensor = model(X_tensor)
    yhat_np = yhat_tensor.cpu().numpy()

# Step 5: Build forecast_df
records = []
for i, (uid, row) in enumerate(df.iterrows()):
    y_true = row["y"]
    yhat_scaled = yhat_np.flatten()  # scalers[uid].inverse_transform(yhat_np[i].reshape(-1, 1))
    ds_vals = row["ds"]
    for ds_i, y_i, yhat_i in zip(ds_vals, y_true, yhat_scaled):
        records.append({"unique_id": uid, "ds": ds_i, "y_scaled": y_i, "yhat_scaled": yhat_i})

forecast_df = pd.DataFrame(records)
forecast_df.to_parquet("data/intermediate/m5_scaled_forecast.parquet")

forecast_df

df
